+++ src/maintenance/__init__.py
@@ -1,0 +1,1 @@
+"""Background maintenance tasks (sweepers, reconciliations, etc.)."""